add forkphorus html file
This is different from what I was originally using, I was originally using the itch.io html embed because that's what I could. Thankfully I was able to get garbomuffin to add a way to grab the assets from an extracted sb3 file (the files are not inside an sb3 file), so I could use my existing project.

A bonus for using this new forkphorus packager option is that the forkphorus option doesn't lag, and I don't have to rely on an external website to keep the page working.